backend/onboarding_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.2
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
websockets==12.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
psutil==5.9.6
